refactor(core): extract Candle RoPE and Conformer blocks
Move create_rope_table / apply_rotary_pos_emb into
conformer/rope.rs and the subsampling / MHSA / FFN / conv
blocks into conformer/blocks.rs. Public ConformerEncoder
paths are unchanged.
